Move shift_accept_reserved & shift_reject_reserved endpoints to ShiftController
